comments_service/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-filter==25.1
djangorestframework==3.15.2
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.9
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
